Changed livewire all properties component
Now it only fetches first image from the media library. This will not
work in the future if I decide to implement a featured photo property,
that will essentially be just a main photo for the property. Just need
to note that down somewhere ... Still need to work on performance (I am
aware that I spelled this wrong so far, sry) issues on pages that don't
contain livewire components or my pc is just slow, not really sure.. But
I definitely need to fix the show.blade that has a full page component.
Will need to split those to two or even less so that not all of them
load whenever I sort properties... Look up the dispatch() function in
Events section of Livewire documentation.