Removed timestamps from attribute list that I send to the view, not its extension of admin/agent/{id} because lets say that agent will have many properties
as the summary says. Not really sure if this is the best way of doing
this, but seem like it might be. Idk if I could ass the user to the
property list in that I create, but even then that is only in show
action, data already traveled to the action