Added delete option to properties listen on agent-property-table livewire component
now it all works, regarding the delete functionality of properties on
admin/property/index and agent/show/{id} blades. Will add a button on
admin/property/show/{id} blade and need to add button to delete agent.
And for now I will work on also just deleting all of his properites,
later on when I work on update functionalities, I will work on ability
to give agents properites to someone else, and then, to delete the
agent. This is all work for tommorow and tonight. As the summary says