Changed the redirect function from back() to admin-properties route at the end of PropertyController/destroy() action
reason for that is, lets say I deleted the property on page X and it is
the only property being displayed on that page. With previous
redirect()->back() it sent admin back to page X which now display no
properites message, which is not good UX. So now, admin will always be
redirected to page 1.