Check:Did shit ton of changes here, removed listener and added comments/loggers to debug this stuff. Some more work to be done, but after this fix.
Found out why does the error happen. Somewhere, component view tries to
query again the photo with deleted photo, because its not computed as it
should be, and that re-query that shouldn't exist is causin me issue.
That is the reason why it doesn't appear when photo is deleted, because
I don't even delete the new photo, there is nothign to query as well, so
it just doesn't fail.... Fix the requery, and its solved