Started working on issue if certain property chart query doesn't have any available data, it shows empty div
which looks ugly. Tried for a hour to add some kind of placeholder to
the chart div instead of the chart, but it doesn't work. Giving up on
that, I will seed it with data and make sure there is no empty queries.
Sorry not sorry, wasting a ton of time on case that will never see live
server.