edited the if statement on agent-property-table that displays color based on lease_duration value.
Now it works as normal. There was issue because at first lease is
nullable, in somewhere along this project I started assigning 0 to it,
instead of 'null' if property was on sale. And there was a lot of
different color bugs not displaying colors properly, because of me just
checking if it was null, not null or 0, as it should because I started
adding 0 to have same meaning as null. Now it does. idc