Edited propertyResultList livewire component
Before there was issue regarding the re-order of listed properties,
which caused all photos from all loaded properties to be fetched. Which
was too much. Previous 4 commits and this are solution to that.

TL;DR made the properties in this component computed, and removed
querying from all other controllers that open properties.index blade,
their only purpose now is to provide filters property to this components
so it can query properties form DB with those filters. and it seems to
work like dandy. :D
This was a really weird bug, and an old one, since I dind't bother to
waste too much time on it before, since I wasn't really comfortable
working with #[Computed] properties. Now it seems to work.

I left the link at the bottom of php file that leads to Livewires
documentation where I got the idea to fix this issue this way using
Computed properites. 07.01.2025.