Edited PropertyObserver
Added if statement in propertyObserver action created and updated to
work properly with properties of action model which require Auth::user
to be present. Since we want these to work with seeders as well to fill
data for charts I added ternary operator to check for auth user if none
is present just set admin as user_id and seed the data.